requirements.txt
python-dotenv
boto3
langchain==1.2.18
langchain-aws==1.4.6
langgraph==1.1.10
langsmith==0.8.3
fastapi==0.136.1
uvicorn==0.46.0